+++ embedded-hal-bus/CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-...
+### Added
+- i2c: add bus sharing implementations.
 
 ## [v0.1.0-alpha.1] - 2022-09-28